legacy: webuser usage cleanup
