+++ invenio/legacy/websession/templates.py
@@ -1398,189 +1398,6 @@ class Template:
                 }
         return out
 
-    def tmpl_create_useractivities_menu(self, ln, selected, url_referer, guest, username, submitter, referee, admin, usebaskets, usemessages, usealerts, usegroups, useloans, usestats, usecomments):
-        """
-        Returns the main navigation menu with actions based on user's
-        priviledges
-
-        @param ln:          The language to display the interface in
-        @type ln:           string
-        @param selected:    If the menu is currently selected
-        @type selected:     boolean
-        @param url_referer: URL of the page being displayed
-        @type url_referer:  string
-        @param guest:       If the user is guest
-        @type guest:        string
-        @param username:    The username (nickname or email)
-        @type username:     string
-        @param submitter:   If the user is submitter
-        @type submitter:    boolean
-        @param referee:     If the user is referee
-        @type referee:      boolean
-        @param admin:       If the user is admin
-        @type admin:        boolean
-        @param usebaskets:  If baskets are enabled for the user
-        @type usebaskets:   boolean
-        @param usemessages: If messages are enabled for the user
-        @type usemessages:  boolean
-        @param usealerts:   If alerts are enabled for the user
-        @type usealerts:    boolean
-        @param usegroups:   If groups are enabled for the user
-        @type usegroups:    boolean
-        @param useloans:    If loans are enabled for the user
-        @type useloans:     boolean
-        @param usestats:    If stats are enabled for the user
-        @type usestats:     boolean
-        @param usecomments: If comments are enabled for the user
-        @type usecomments:  boolean
-        @return: html menu of the user activities
-        @rtype: string
-        """
-        # load the right message language
-        _ = gettext_set_language(ln)
-
-        out = '''<div class="hassubmenu%(on)s">
-        <a hreflang="en" class="header%(selected)s" href="%(CFG_SITE_SECURE_URL)s/youraccount/display?ln=%(ln)s">%(personalize)s</a>
-        <ul class="subsubmenu">''' % {
-                'CFG_SITE_SECURE_URL' : CFG_SITE_SECURE_URL,
-                'ln' : ln,
-                'personalize': _("Personalize"),
-                'on': selected and " on" or '',
-                'selected': selected and "selected" or ''
-                }
-        if not guest:
-            out += '<li><a href="%(CFG_SITE_SECURE_URL)s/youraccount/display?ln=%(ln)s">%(account)s</a></li>'  % {
-                'CFG_SITE_SECURE_URL' : CFG_SITE_SECURE_URL,
-                'ln' : ln,
-                'account' : _('Your account')
-                }
-        if usealerts or guest:
-            out += '<li><a href="%(CFG_SITE_SECURE_URL)s/youralerts/list?ln=%(ln)s">%(alerts)s</a></li>'  % {
-                'CFG_SITE_SECURE_URL' : CFG_SITE_SECURE_URL,
-                'ln' : ln,
-                'alerts' : _('Your alerts')
-                }
-        if referee:
-            out += '<li><a href="%(CFG_SITE_SECURE_URL)s/yourapprovals.py?ln=%(ln)s">%(approvals)s</a></li>'  % {
-                'CFG_SITE_SECURE_URL' : CFG_SITE_SECURE_URL,
-                'ln' : ln,
-                'approvals' : _('Your approvals')
-                }
-        if usebaskets or guest:
-            out += '<li><a href="%(CFG_SITE_SECURE_URL)s/yourbaskets/display?ln=%(ln)s">%(baskets)s</a></li>'  % {
-                'CFG_SITE_SECURE_URL' : CFG_SITE_SECURE_URL,
-                'ln' : ln,
-                'baskets' : _('Your baskets')
-                }
-        if usecomments:
-            out += '<li><a href="%(CFG_SITE_SECURE_URL)s/yourcomments?ln=%(ln)s">%(comments)s</a></li>'  % {
-                'CFG_SITE_SECURE_URL' : CFG_SITE_SECURE_URL,
-                'ln' : ln,
-                'comments' : _('Your comments')
-                }
-        if usegroups:
-            out += '<li><a href="%(CFG_SITE_SECURE_URL)s/yourgroups/display?ln=%(ln)s">%(groups)s</a></li>'  % {
-                'CFG_SITE_SECURE_URL' : CFG_SITE_SECURE_URL,
-                'ln' : ln,
-                'groups' : _('Your groups')
-                }
-        if useloans:
-            out += '<li><a href="%(CFG_SITE_SECURE_URL)s/yourloans/display?ln=%(ln)s">%(loans)s</a></li>'  % {
-                'CFG_SITE_SECURE_URL' : CFG_SITE_SECURE_URL,
-                'ln' : ln,
-                'loans' : _('Your loans')
-                }
-        if usemessages:
-            out += '<li><a href="%(CFG_SITE_SECURE_URL)s/yourmessages/display?ln=%(ln)s">%(messages)s</a></li>'  % {
-                'CFG_SITE_SECURE_URL' : CFG_SITE_SECURE_URL,
-                'ln' : ln,
-                'messages' : _('Your messages')
-                }
-        if submitter:
-            out += '<li><a href="%(CFG_SITE_SECURE_URL)s/yoursubmissions.py?ln=%(ln)s">%(submissions)s</a></li>'  % {
-                'CFG_SITE_SECURE_URL' : CFG_SITE_SECURE_URL,
-                'ln' : ln,
-                'submissions' : _('Your submissions')
-                }
-        if usealerts or guest:
-            out += '<li><a href="%(CFG_SITE_SECURE_URL)s/youralerts/display?ln=%(ln)s">%(searches)s</a></li>'  % {
-                'CFG_SITE_SECURE_URL' : CFG_SITE_SECURE_URL,
-                'ln' : ln,
-                'searches' : _('Your searches')
-                }
-        out += '</ul></div>'
-        return out
-
-    def tmpl_create_adminactivities_menu(self, ln, selected, url_referer, guest, username, submitter, referee, admin, usebaskets, usemessages, usealerts, usegroups, useloans, usestats, activities):
-        """
-        Returns the main navigation menu with actions based on user's
-        priviledges
-
-        @param ln:          The language to display the interface in
-        @type ln:           string
-        @param selected:    If the menu is currently selected
-        @type selected:     boolean
-        @param url_referer: URL of the page being displayed
-        @type url_referer:  string
-        @param guest:       If the user is guest
-        @type guest:        string
-        @param username:    The username (nickname or email)
-        @type username:     string
-        @param submitter:   If the user is submitter
-        @type submitter:    boolean
-        @param referee:     If the user is referee
-        @type referee:      boolean
-        @param admin:       If the user is admin
-        @type admin:        boolean
-        @param usebaskets:  If baskets are enabled for the user
-        @type usebaskets:   boolean
-        @param usemessages: If messages are enabled for the user
-        @type usemessages:  boolean
-        @param usealerts:   If alerts are enabled for the user
-        @type usealerts:    boolean
-        @param usegroups:   If groups are enabled for the user
-        @type usegroups:    boolean
-        @param useloans:    If loans are enabled for the user
-        @type useloans:     boolean
-        @param usestats:    If stats are enabled for the user
-        @type usestats:     boolean
-        @param activities: dictionary of admin activities
-        @rtype activities: dict
-        @return: html menu of the user activities
-        @rtype: string
-        """
-        # load the right message language
-        _ = gettext_set_language(ln)
-
-        out = ''
-        if activities:
-            out += '''<div class="hassubmenu%(on)s">
-            <a hreflang="en" class="header%(selected)s" href="%(CFG_SITE_SECURE_URL)s/youraccount/youradminactivities?ln=%(ln)s">%(admin)s</a>
-            <ul class="subsubmenu">''' % {
-            'CFG_SITE_SECURE_URL' : CFG_SITE_SECURE_URL,
-            'ln' : ln,
-            'admin': _("Administration"),
-            'on': selected and " on" or '',
-            'selected': selected and "selected" or ''
-            }
-
-            for name in sorted(activities.iterkeys()):
-                url = activities[name]
-                out += '<li><a href="%(url)s">%(name)s</a></li>'  % {
-                    'url': url,
-                    'name': name
-                    }
-
-        if usestats:
-            out += """<li><a href="%(CFG_SITE_URL)s/stats/?ln=%(ln)s">%(stats)s</a></li>""" % {
-                'CFG_SITE_URL' : CFG_SITE_URL,
-                'ln' : ln,
-                'stats' : _("Statistics"),
-                }
-
-            out += '</ul></div>'
-        return out
-
     def tmpl_warning(self, warnings, ln=CFG_SITE_LANG):
         """
         Display len(warnings) warning fields
